+++ webservice/src/main/java/de/fau/cs/osr/amos/asepart/messages/Credentials.java
@@ -1,0 +1,29 @@
+package de.fau.cs.osr.amos.asepart.messages;
+
+import java.io.Serializable;
+
+public class Credentials implements Serializable
+{
+    private String username;
+    private String password;
+
+    public String getUsername()
+    {
+        return username;
+    }
+
+    public void setUsername(String username)
+    {
+        this.username = username;
+    }
+
+    public String getPassword()
+    {
+        return password;
+    }
+
+    public void setPassword(String password)
+    {
+        this.password = password;
+    }
+}
